requirements.txt
Django==5.1
django-admin-rangefilter==0.13.1
django-debug-toolbar==4.4.6
Faker==26.2.0
python-dotenv==1.0.1
